imageprocessor/requirements.txt
appdirs==1.4.4
astunparse==1.6.3
async-timeout==3.0.1
blinker==1.4
brotlipy==0.7.0
certifi==2020.11.8
click==7.1.2
cycler==0.10.0
distlib==0.3.1
filelock==3.0.12
gast==0.3.3
idna-ssl==1.1.0
importlib-metadata==1.7.0
importlib-resources==3.0.0
Keras-Preprocessing==1.1.0
kiwisolver==1.3.1
matplotlib==3.3.2
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
numpy==1.19.4
oauthlib==3.1.0
opt-einsum==3.1.0
pandas==1.1.4
Pillow==8.0.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
requests-oauthlib==1.3.0
six==1.15.0
tensorboard-plugin-wit==1.6.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
virtualenv==20.0.30
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
